% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabetize.R
\name{alpha_sort}
\alias{alpha_sort}
\title{Sort by Alphabet}
\usage{
alpha_sort(text)
}
\arguments{
\item{text}{String, user text selection}
}
\value{
sorted text
}
\description{
Helper function for alphabetize addin. \cr
Given active text selection, the function parses the text and
re-order them by the alphabet.
}
\examples{
cat(
  alpha_sort(
    'library(tidyr)
     library(stringr)
     library(dplyr)'))

}
