% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{code_counter}
\alias{code_counter}
\title{Code Lines Counter}
\usage{
code_counter(text)
}
\arguments{
\item{text}{String / list of Strings}
}
\value{
Integer, code lines count
}
\description{
This function is being used internally by \code{code_stats}.
}
\examples{
\dontrun{
code_counter("# comment")
code_counter("2 * 2")
code_counter(c("(2 * 2)", ("2 * 2")))
}
}
