% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{code_stats}
\alias{code_stats}
\title{Code Statistics}
\usage{
code_stats(text)
}
\arguments{
\item{text}{String}
}
\value{
String, code stats
}
\description{
Counts code lines, comment lines, blank lines and code to comments ratio.
}
\examples{
cat(
  code_stats("# R calculations
             2 * 2 / 4
             (4 * 4) / 2
             rnorm(2)"))

}
