% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify_io}
\alias{notify_io}
\title{Notify on IO components in the code}
\usage{
notify_io(text)
}
\arguments{
\item{text}{String}
}
\value{
String
}
\description{
Notify on IO components in the code
}
\examples{
cat(
  notify_io(
    'df <- read.csv("my_data.csv")
    write.csv(df, "your_data.csv")
    write_variable <- "will not be mixed up with write function"'))
}
