% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify_libraries}
\alias{notify_libraries}
\title{Detect and Display Libraries Used Throughout the Script.}
\usage{
notify_libraries(text)
}
\arguments{
\item{text}{String}
}
\value{
String
}
\description{
Detect and Display Libraries Used Throughout the Script.
}
\examples{
cat(
  notify_libraries(
    'require(dplyr)
    library(magrittr)
    stringr::str_replace("1", "ONE")'))

}
