% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_functions}
\alias{list_functions}
\title{List functions in current script}
\usage{
list_functions(text)
}
\arguments{
\item{text}{String}
}
\value{
commented list of functions
}
\description{
Any functions appearing in the input text
are identified and their names. Names are then sorted
alphabetically and outputted as a commented text in a
"function per line" fashion.
}
